refactor(library/Sequence): moves "Base64/Conformance/exports.ts" to "Base64/Conformance/ensure.ts"
