+++ src/library/Sequence/Base64/Conformance/index.ts
@@ -1,1 +1,1 @@
-export * as Sequence_Base64_Conformance from './exports.ts';
+export * as Sequence_Base64_Conformance from './ensure.ts';
